requirements.txt
django>=5.0
djangorestframework>=3.14
pymongo>=4.6
